+++ packages/network-controller/src/create-network-client.ts
@@ -34,7 +34,7 @@ import type {
   NetworkClientId,
   NetworkControllerMessenger,
 } from './NetworkController';
-import type { RpcServiceOptions } from './rpc-service/rpc-service';
+import type { RpcServiceOptionsWithDefaults } from './rpc-service/rpc-service';
 import {
   isConnectionError,
   isConnectionResetError,
@@ -146,7 +146,7 @@ export function createNetworkClient({
   configuration: NetworkClientConfiguration;
   getRpcServiceOptions: (
     rpcEndpointUrl: string,
-  ) => Omit<RpcServiceOptions, 'failoverService' | 'endpointUrl'>;
+  ) => RpcServiceOptionsWithDefaults;
   getBlockTrackerOptions: (
     rpcEndpointUrl: string,
   ) => Omit<PollingBlockTrackerOptions, 'provider'>;
@@ -254,7 +254,7 @@ function createRpcServiceChain({
   configuration: NetworkClientConfiguration;
   getRpcServiceOptions: (
     rpcEndpointUrl: string,
-  ) => Omit<RpcServiceOptions, 'failoverService' | 'endpointUrl'>;
+  ) => RpcServiceOptionsWithDefaults;
   messenger: NetworkControllerMessenger;
   isRpcFailoverEnabled: boolean;
   logger?: Logger;
@@ -271,10 +271,12 @@ function createRpcServiceChain({
   };
 
   const rpcServiceConfigurations = availableEndpointUrls.map((endpointUrl) => ({
+    fetch: globalThis.fetch.bind(globalThis),
+    btoa: globalThis.btoa.bind(globalThis),
+    isOffline,
     ...getRpcServiceOptions(endpointUrl),
     endpointUrl,
     logger,
-    isOffline,
   }));
 
   /**